VFS does not keep a path anymore
Previously, a VFS had the notion of a 'path' but it really makes
little sense to have a VFS limited to it. Instead a VFS provides
a consistent interface to the underlying filesystem module (so
far, local & gdrive exist.) FileTree now takes a 'path' parameter
to operate on and keeps no state in the VFS object.